+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Change the upgrade progress bars to be more accurate.
 - Change the upgrade to only upgrade Mac drivers if Apple hardware is detected.
 - Change the first-time setup to install drivers from the [extra] repository instead of [community] now that they have been merged.
+- Add Pacman support in the Discover package manager.
 
 ## 3.2.1 - The SteamOS 3.4 Update